Moving decoder
