+++ packages/langium/src/grammar/lsp/grammar-semantic-tokens.ts
@@ -4,11 +4,10 @@
  * terms of the MIT License, which is available in the project root.
  ******************************************************************************/
 
-import { SemanticTokens, SemanticTokenTypes } from 'vscode-languageserver';
+import { SemanticTokenTypes } from 'vscode-languageserver';
 import { AstNode } from '../../syntax-tree';
-import { AbstractSemanticTokenProvider, AllSemanticTokenTypes, SemanticTokenAcceptor } from '../../lsp/semantic-token-provider';
+import { AbstractSemanticTokenProvider, SemanticTokenAcceptor } from '../../lsp/semantic-token-provider';
 import { isAction, isAssignment, isAtomType, isParameter, isParameterReference, isReturnType } from '../generated/ast';
-import { LangiumDocument } from '../../workspace/documents';
 
 export class LangiumGrammarSemanticTokenProvider extends AbstractSemanticTokenProvider {
 
@@ -56,44 +55,4 @@ export class LangiumGrammarSemanticTokenProvider extends AbstractSemanticTokenPr
         }
     }
 
-}
-
-export namespace SemanticTokensDecoder {
-    export interface DecodedSemanticToken {
-        offset: number;
-        tokenType: SemanticTokenTypes;
-        tokenModifiers: number;
-        text: string;
-    }
-
-    export function decode<T extends AstNode = AstNode>(tokens: SemanticTokens, document: LangiumDocument<T>): DecodedSemanticToken[] {
-        const typeMap = new Map<number, SemanticTokenTypes>();
-        Object.entries(AllSemanticTokenTypes).forEach(([type, index]) => typeMap.set(index, type as SemanticTokenTypes));
-        let line = 0;
-        let character = 0;
-        return sliceIntoChunks(tokens.data, 5).map(t => {
-            line += t[0];
-            if (t[0] !== 0) {
-                character = 0;
-            }
-            character += t[1];
-            const length = t[2];
-            const offset = document.textDocument.offsetAt({ line, character });
-            return {
-                offset,
-                tokenType: typeMap.get(t[3])!,
-                tokenModifiers: t[4],
-                text: document.textDocument.getText({ start: { line, character }, end: { line, character: character + length } })
-            } as DecodedSemanticToken;
-        });
-    }
-
-    function sliceIntoChunks<T>(arr: T[], chunkSize: number) {
-        const res = [];
-        for (let i = 0; i < arr.length; i += chunkSize) {
-            const chunk = arr.slice(i, i + chunkSize);
-            res.push(chunk);
-        }
-        return res;
-    }
 }
